Update platformio config for new version

+++ software/power_circuit/src/main.cpp
@@ -1,10 +1,12 @@
-/*  
- *  power_circuit.ino
- *  Program to control power to a RbPi 2 with safe shutdown.  
+/*
+ *  power_circuit.cpp
+ *  Program to control power to a RPi with safe shutdown. 
  *
- *  Created on: Dec 15, 2015
+ *  Created on: Feb 27, 2019
  *  Author: Johan Wermensjö
  */
+ 
+#include <Arduino.h>
 
 // Definitions.
 //#define AUTO_ON
@@ -36,7 +38,7 @@
 // Input states.
 #define POWER_BUTTON_UP LOW
 #define POWER_BUTTON_DOWN HIGH
-#define POWER_CTRL_READY_THRESHOLD (uint16_t) 256 // Corresponds to 1.25V
+#define POWER_CTRL_READY_THRESHOLD (int) 256 // Corresponds to 1.25V
 
 // Functions.
 static void waitForButtonState(uint8_t state);
@@ -53,47 +55,47 @@ void setup() {
 	pinMode(POWER_BUTTON_LED_OUTPUT_PIN, OUTPUT);
 	pinMode(POWER_SWITCH_OUTPUT_PIN, OUTPUT);
 	pinMode(POWER_BUTTON_INPUT_PIN, INPUT);
-	
+
 	// Initialize power state.
 	#ifdef AUTO_ON
 		enablePower(true);
 	#else
 		enablePower(false);
 	#endif
-	
+
 	// Initialize no shutdown alert.
 	digitalWrite(POWER_CTRL_OUTPUT_PIN, POWER_ALERT_OFF);
 }
 
 void loop() {
-	
+
 	// Make sure button is released between state changes.
 	waitForButtonState(POWER_BUTTON_UP);
 
 	// Wait for button press to toggle power.
 	waitForButtonState(POWER_BUTTON_DOWN);
-	
+
 	if (powerOn) {
 		/*****************************/
 		/** Power on - Turning off. **/
 
 		// Alert to begin shutdown procedure.
 		digitalWrite(POWER_CTRL_OUTPUT_PIN, POWER_ALERT_ON);
-	
+
 		// Initiate button timeout.
 		uint16_t powerButtonnHeldTime = 0;
 		uint16_t ledToggleTime = 0;
 		uint16_t ctrlReadyTime = 0;
 		bool ledState = true;
 		bool ctrlReadyReceived = false;
-	
+
 		while (powerButtonnHeldTime < POWER_BUTTON_HOLD_TIME) {
 			// Wait to next check.
 			delay(POWER_BUTTON_POLL_DELAY);
-			
+
 			// Add delay to toggle time.
 			ledToggleTime += POWER_BUTTON_POLL_DELAY;
-	
+
 			// Toggle led after the set time.
 			if (ledToggleTime >= POWER_BUTTON_LED_BLINK_DELAY ||
 					(ctrlReadyReceived && ledToggleTime >= POWER_BUTTON_LED_FAST_BLINK_DELAY)) {
@@ -105,11 +107,11 @@ void loop() {
 				ledState = !ledState;
 				ledToggleTime = 0;
 			}
-	
+
 			// Listen for ready signal.
 			if (isControlReady()) {
 				ctrlReadyTime += POWER_BUTTON_POLL_DELAY;
-				
+
 				// Set ready signal as recived when held for required time.
 				if (ctrlReadyTime >= POWER_CTRL_READY_HOLD_TIME) {
 					ctrlReadyReceived = true;
@@ -124,7 +126,7 @@ void loop() {
 				ctrlReadyReceived = false;
 				ctrlReadyTime = 0;
 			}
-	
+
 			// If button is held for timeout then force shutdown.
 			if (digitalRead(POWER_BUTTON_INPUT_PIN) == POWER_BUTTON_DOWN) {
 				// Button is still pressed, add delay to counter.
@@ -134,17 +136,17 @@ void loop() {
 				powerButtonnHeldTime = 0;
 			}
 		}
-		
+
 		// Turn off power.
 		enablePower(false);
-		
+
 		// Turn off alert.
 		digitalWrite(POWER_CTRL_OUTPUT_PIN, POWER_ALERT_OFF);
-		
+
 	} else {
 		/*****************************/
 		/** Power off - Turning on. **/
-		
+
 		// Turn on power.
 		enablePower(true);
 	}
